zhanggh  修改绣花物料及仓库管理

+++ trfc-service/src/main/java/com/tianrui/service/mapper/basicFile/nc/MaterielManageMapper.java
@@ -24,7 +24,7 @@ public interface MaterielManageMapper {
 
     int updateByPrimaryKey(MaterielManage record);
     
-   	int findMaterielManagePageCount(MaterielManageReq req);
+    long findMaterielManagePageCount(MaterielManageReq req);
     
     List<MaterielManage> findMaterielManagePage(MaterielManageReq req);
